Fix unit tests from refactoring (task #4333)
